complete redesign
In an effort to retain timestamps a new approach was taken where xslt
was used to produce an intermediate representation which could be
transformed by dateconv before using xslt once again to write the
files.

The major change is user-defined titles are no longer used, instead the
title found in the feed is substituted (or the domain if lacking).

As a result the filesystem layout has been changed to keep curl
artefacts and the database separate.