Return to POSIX awk usage and fix off-by-one error
This change uses substr() and a for loop to iterate over each byte in
the input instead of splitting the bytes into an array using unspecified
behaviour with split(), from awk(1p):

    > The effect of a null string as the value of fs is unspecified.

As a consequence m4 needs to have its substr macro disabled as otherwise
it would conflict with its use in awk.

Finally I addressed the issue that my loop ranged from [0, 255) instead
of [0, 256).  Fortunately the chance that 0xff would appear in a real
example is slim and, at least for me, caused no issues.