webfeeds: remove extraneous url in error condition
curl used to be instructed to print the url as well as an error message
when encountering an error.  The url was used to determine if the
failure left behind a partial file and to remove it.

curl then added --remove-on-error which I switched to using in
3457f81f8a171f63dace311d45c6524de5898149, however, I had forgotten to
remove the url printing from the error message.